Fix issue where rust lsp commands weren't working
The problem was that rust-tools was being enabled and nuking the buffer
key mappings that were previously setup. So now I move the key mappings
out of the lsp into just global key mappings.

I could have tried to tie them to specific languages are what not. But
at this point I am pretty much using the LSP commands for everything or
dealing with a language where they aren't supported. In which case I
don't want to repurpose the keyboard commands. Therefore it should be
fine to have them global.

I also added vsnip and cmp-vsnip and configured it as the snippet
manager so that autocompletion would work properly with it.

Beyond that I cleaned up some of the formatting and replaced a bunch of
tab characters with two spaces.