Disable clojure plugin & cleanup Rust config
My intention with this change is to disable the clojure socket repl
plugin and to cleanup my Rust configuration and see if I can get it and
LSP to work a little better. I am disabling the clojure socket repl
because I am working in clojure anymore and it is an unnecessary cost
at bootup. My rust LSP and snippet setup isn't performing well and seems
to get into broken states every once and a while especially when trying
to complete snippets so I am trying to adjust the configurations to see
if it helps.