Get rid of trouble & update lsp keycommands
Get rid of trouble & update lsp keycommands so that all the key commands
use the new non-deprecated api. I have found that working with these key
commands it provides everything I want. It lets me see a floating window
with the detailed diagnostic as well as it lets me send the diagnostics
to the quickfix list to process them one by one, etc.