Add ensure_installed grammers to tree sitter conf
Add ensure_installed grammers to tree sitter conf so that when I launch
neovim it will make sure that those grammers are installed. This just
makes it easier when I setup my neovim setup on a new machine as I don't
have to remember to install each of these grammers manually.

<!-- ps-id: f545f2ad-abbc-428d-b8f2-ff41c0e0d55f -->