Remove the Ctrl-c mapping to esc
Remove the Ctrl-c mapping to esc because in weird situations it doesn't
work exactly the way I would want and it is a bad habbit because it
requires two key presses, one from each hand, since I am in dvorak. I
just need to retrain myself to use the esc key now that I have my new
keyboard layout and keyboard as it should be more efficient in the long
run.

<!-- ps-id: cbfd648e-1166-4da5-80f0-dda3870e494b -->