Temp disable eslint configuration
Temp disable eslint configuration, I believe because I was in a project
where eslint was fighting the auto formatting from the lsp and I didn't
want to take the time to figure out the two tools fighting each other in
that project. So it is commented out for the time being so that when and
if I decide to bring it back it will be here as a reference.

<!-- ps-id: f27b19b2-e9d2-4df1-923c-9e64e0852dee -->