Address treesitter dart indent perf problems
Address treesitter dart indent perf problems so that I could indent in
dart files without huge multi-second delays.

In this I ultimately addressed this issue by disabling treesitter indent
for dart in the configuration.

However, prior to that I addressed it by locking the treesitter plugin
to a previous version. This approach had a number of other
complications. But, I wanted to keep the sha around just in case I
needed to go back to that at some point. So I left it in commented out.

Another approach I used was to lock the parser at a specific previous
version. This seemed to work kind of but didn't completely solve the
issue. But, I kept the reference in as a comment along with the details
of the associated GitHub issue.

Ultimately I decided to just disable treesitter indent for dart. At some
point in the future I should probably try enabling it again and cleaning
all of this up once they fix the issue in the parser.

<!-- ps-id: 958ceba8-253f-4c72-a174-a81cf9a0b8e1 -->