Migrate to lazy.nvim
Migrate to lazy.nvim largely because I needed to just clean up my
dotnvim stuff a bit and figured I might as well get the benefits of lazy
loading plugins as well as the version locking, etc.

So this is the result of at least initially migrating my dotnvim over to
use lazy.nvim.