Add keybindings for LSP to rust-tools
Add keybindings for LSP to rust-tools so that the keybindings would
actually work in Rust files.

This is a less than ideal approach to solving this that I just threw in
here quickly to get back to working, as it duplicates the keybindings. I
should likely extract it out further up as a dependency or something and
then use it in both places. Either that or move the rust-tools setup
into the lsp plugin configuration function somehow where the local key
bindings function already exists.

<!-- ps-id: 6af810d5-290f-42e6-a234-3c72070c705a -->